+++ src/builtins/export.c
@@ -67,7 +67,7 @@ int	check_existence(char *var, char **envp)
 	return (0);
 }
 
-int	builtin_export(int argc, char **argv, t_envp *envp_status, int *fds)
+int	builtin_export(size_t argc, char **argv, t_envp *envp_status, int *fds)
 {
 	int	error;
 
@@ -76,7 +76,7 @@ int	builtin_export(int argc, char **argv, t_envp *envp_status, int *fds)
 		error = (int)export_only(*envp_status, fds[WRITE]);
 		if (error == 2)
 			ft_perror("minishell: export: malloc(): \
-			failed memory allocation\n");
+failed memory allocation\n");
 		return (error);
 	}
 	error = (int)export_to_envp(&argv[1], envp_status);
